discord-bots/ecosystem/requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
pytz>=2021.3
pandas>=1.3.0
numpy>=1.21.0
yfinance>=0.1.70
beautifulsoup4>=4.10.0
requests>=2.26.0
python-dateutil>=2.8.2
